Working version for Hannah
count_eggs.ipynb contains a workflow of processing egg images to count eggs. Far from perfect and may need more sophisticated procedure but this is as far as I was able to automate it. The input images could be stored in the subdirectory within the folder or uploaded from anywhere on the machine. Need to update readme.